Examples/Python Learnings/Django_Projects/website/companies/serializers.py
from rest_framework import serializers
from .models import Stocks


class StocksSerializer(serializers.ModelSerializer):
    class Meta:
        model = Stocks
        # fields = ('ticker', 'volume')
        fields = '__all__'
