Examples/Python Learnings/Django_Projects/requirements.txt
asgiref==3.2.5
asn1crypto==0.24.0
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.1.0
backports.functools-lru-cache==1.5
bcrypt==3.1.7
caniusepython3==7.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
distlib==0.2.9.post0
Django==3.0.4
django-crispy-forms==1.9.0
djangorestframework==3.11.0
idna==2.8
importlib-metadata==0.19
isort==4.3.21
kazoo==2.5.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.2
lxml==4.4.1
matplotlib==3.1.1
mccabe==0.6.1
more-itertools==7.2.0
numpy==1.17.0
packaging==19.1
pandas==0.25.1
paramiko==2.6.0
Pillow==7.0.0
pluggy==0.12.0
py==1.8.0
pycparser==2.19
pylint==2.4.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
pytest==4.5.0
pytest-timeout==1.3.3
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
scp==0.13.2
six==1.12.0
sqlparse==0.3.1
typed-ast==1.4.0
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.2
xmlformatter==0.1.1
xmltodict==0.12.0
zipp==0.5.2
